backend/onego/views.py
# This file is now for truly global or cross-domain API endpoints only.
# For single-domain logic, use each app's own views.py and serializers.py!

from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated

class SettingsView(APIView):
    permission_classes = [IsAuthenticated]
    def get(self, request):
        # Example: return user/theme/settings
        return Response({"theme": "pink-blue-black"})
    def put(self, request):
        # Extend: save theme/settings for user
        return Response({"message": "Settings updated"})