backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
celery
redis
django-environ
channels
channels-redis
razorpay
firebase-admin
twilio
gunicorn==20.1.0           # or whatever the latest
django-extensions>=4.1.0,<5.0

